pyproject.toml
[project]
name = "extract-from-pdf"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pymupdf>=1.24.0",
    "requests>=2.31.0",
    "pyyaml>=6.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[dependency-groups]
dev = [
    "mypy>=1.20.2",
    "types-pyyaml>=6.0.12.20260408",
    "types-requests>=2.33.0.20260503",
]
